Centralize types in new index.ts, update imports across project
