+++ apps/roam/src/components/DiscourseNodeSearchMenu.tsx
@@ -21,7 +21,8 @@ import updateBlock from "roamjs-components/writes/updateBlock";
 import posthog from "posthog-js";
 import { getCoordsFromTextarea } from "roamjs-components/components/CursorMenu";
 import { OnloadArgs } from "roamjs-components/types";
-import getDiscourseNodes, { DiscourseNode } from "~/utils/getDiscourseNodes";
+import getDiscourseNodes from "~/utils/getDiscourseNodes";
+import { DiscourseNode } from "~/types/index";
 import getDiscourseNodeFormatExpression from "~/utils/getDiscourseNodeFormatExpression";
 import { escapeCljString } from "~/utils/formatUtils";
 import { Result } from "~/utils/types";
