+++ apps/roam/src/utils/discourseNodeFormatToDatalog.ts
@@ -1,7 +1,7 @@
 import { DatalogClause } from "roamjs-components/types/native";
 import conditionToDatalog from "./conditionToDatalog";
 import getDiscourseNodeFormatExpression from "./getDiscourseNodeFormatExpression";
-import type { DiscourseNode } from "./getDiscourseNodes";
+import type { DiscourseNode } from "~/types/index";
 import replaceDatalogVariables from "./replaceDatalogVariables";
 
 const discourseNodeFormatToDatalog = ({
