+++ apps/roam/src/utils/getDiscourseRelations.ts
@@ -7,16 +7,7 @@ import getSettingValueFromTree from "roamjs-components/util/getSettingValueFromT
 import toFlexRegex from "roamjs-components/util/toFlexRegex";
 import DEFAULT_RELATION_VALUES from "~/data/defaultDiscourseRelations";
 import discourseConfigRef from "./discourseConfigRef";
-
-export type Triple = readonly [string, string, string];
-export type DiscourseRelation = {
-  triples: Triple[];
-  id: string;
-  label: string;
-  source: string;
-  destination: string;
-  complement: string;
-};
+import { Triple, DiscourseRelation } from "~/types/index";
 
 const matchNodeText = (keyword: string) => {
   return (node: RoamBasicNode | TextNode) =>
