+++ apps/roam/src/utils/parseResultSettings.ts
@@ -5,7 +5,7 @@ import getSettingIntFromTree from "roamjs-components/util/getSettingIntFromTree"
 import getSubTree from "roamjs-components/util/getSubTree";
 import toFlexRegex from "roamjs-components/util/toFlexRegex";
 import { StoredFilters } from "~/components/settings/DefaultFilters";
-import { Column } from "./types";
+import { Column } from "~/types/index";
 import getSettingValueFromTree from "roamjs-components/util/getSettingValueFromTree";
 import getSettingValuesFromTree from "roamjs-components/util/getSettingValuesFromTree";
 import {
